Rename sunxspex fitting module to fitting.
