+++ changelog/133.trivial.rst
@@ -1,0 +1,1 @@
+Rename ``sunxspex_fitting`` module to ``fitting``.
