+++ sunkit_spex/fitting/fitter.py
@@ -36,17 +36,17 @@ from scipy.optimize import minimize
 from astropy.table import Table
 
 from sunkit_spex.logging import get_logger
-from sunkit_spex.sunxspex_fitting.data_loader import LoadSpec
-from sunkit_spex.sunxspex_fitting.instruments import rebin_any_array
-from sunkit_spex.sunxspex_fitting.likelihoods import LogLikelihoods
-from sunkit_spex.sunxspex_fitting.parameter_handler import Parameters, isnumber
-from sunkit_spex.sunxspex_fitting.photon_models_for_fitting import (  # noqa
+from sunkit_spex.fitting.data_loader import LoadSpec
+from sunkit_spex.fitting.instruments import rebin_any_array
+from sunkit_spex.fitting.likelihoods import LogLikelihoods
+from sunkit_spex.fitting.parameter_handler import Parameters, isnumber
+from sunkit_spex.fitting.photon_models_for_fitting import (  # noqa
     defined_photon_models,
     f_vth,
     thick_fn,
     thick_warm,
 )
-from sunkit_spex.sunxspex_fitting.rainbow_text import rainbow_text_lines
+from sunkit_spex.fitting.rainbow_text import rainbow_text_lines
 
 warnings.filterwarnings("ignore", category=RuntimeWarning)
 warnings.filterwarnings("ignore", category=np.VisibleDeprecationWarning)
@@ -66,7 +66,7 @@ class Fitter:
     Parameters
     ----------
     *args : dict
-            Dictionaries for custom data to be passed to `sunkit_spex.sunxspex_fitting.instruments.CustomLoader`.
+            Dictionaries for custom data to be passed to `sunkit_spex.fitting.instruments.CustomLoader`.
             These will be added before any instrument file entries from `pha_file`.
 
     pha_file : string or list of strings
@@ -778,7 +778,7 @@ class Fitter:
             logger.indo(f"Model {function_name} removed.")
         else:
             logger.warning(
-                "Default models imported from sunkit_spex.sunxspex_fitting.photon_models_for_fitting are protected.")
+                "Default models imported from sunkit_spex.fitting.photon_models_for_fitting are protected.")
 
     def add_var(self, overwrite=False, quiet=False, **user_kwarg):
         """ Add user variable to fitting namespace.
