+++ sunkit_spex/fitting/rainbow_text.py
