+++ sunkit_spex/fitting/tests/test_io.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock, patch
 
 import numpy as np
 
-from sunkit_spex.sunxspex_fitting.io import (
+from sunkit_spex.fitting.io import (
     _read_arf,
     _read_pha,
     _read_rhessi_spec_file,
