+++ sunkit_spex/fitting/tests/test_likelihoods.py
@@ -1,7 +1,7 @@
 import numpy as np
 from numpy.testing import assert_allclose
 
-from sunkit_spex.sunxspex_fitting.likelihoods import LogLikelihoods
+from sunkit_spex.fitting.likelihoods import LogLikelihoods
 
 
 def test_cstat():
